server/requirements (1).txt
fastapi
uvicorn
pydantic
pytz
firebase-admin
